Set Java version for maven-compiler-plugin in non Spring Boot modules
In a Spring Boot module, the spring-boot-maven-plugin takes care of
setting Java version when compiling source code, using the Maven
property 'java.version'. When this plugin is not included in the POM
(like in a regular Java library), source and target must be set so that
maven-compiler-plugin knows how to compile source code.